ref: getting load blockchain logic out of relayer
